% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra finds the shortest distance between two nodes.}
\usage{
dijkstra(df, s)
}
\arguments{
\item{df}{the first parameter}

\item{s}{the second parameter}
}
\description{
Dijkstra finds the shortest path to a graph
}
